html2text: add a basic command-line parser.
I've used the same argument-parsing crate as html2term, and used it to
implement a command-line option to set the column width of the
formatted output.

While I'm at it, I've also allowed specifying an input file name as a
positional argument, and the usual -o option to write to an output
file, so that html2text doesn't always have to be used with its
standard I/O handles redirected. However, the default behaviour if
neither of those options is present is to read from standard input and
write to standard output, so this change should be backwards
compatible in the sense that html2text run without arguments should
continue to do what it always has done. (However, running it _with_
arguments will now change behaviour, because the arguments aren't
ignored any more!)

This patch would look nicer if my Rust were good enough to figure out
how to set a variable to either stdin or the result of opening a file,
so that I could have just one call to html2text::from_read, and
similarly on output. But it's not. Sorry about that.