html2term: use the alternate screen.
Previously, pressing 'q' would terminate the viewer immediately,
leaving the next shell prompt to appear wherever the cursor had
happened to be. Now we do all the document display on the alternate
screen, and the old one is automatically restored on exit, so the
display state after html2term finishes is more sensible.