html2term: rework position, scrolling and navigation.
Now, instead of remembering the line at the top of the screen and the
cursor position within the screen, we primarily remember the cursor
position within the _document_, and the cursor position on the screen
is derived from that when it's needed.

This makes it easier to implement the various navigation commands
sensibly: for example, Up and Down now reliably move the cursor one
line up or down the document, and a standard piece of pre-display
clipping ensures that the visible screen rectangle scrolls to keep the
cursor in view. (Previously, pressing Up or Down at the top or bottom
of the screen respectively didn't do anything, because there was no
'else' clause to handle it.)

Also, we now support Home and End to go to the top or bottom of the
document.