Fragment support: handle fragment ids in TextRenderer.
This commit fills in TextRenderer's implementation of
record_frag_start.

To pass fragment ids through the word-wrapping system, I've changed
the TaggedLine type so that instead of holding a vector of
TaggedString, it now holds a vector of an enumerated type which can be
a TaggedString or some other kind of marker that doesn't contain text.
Currently the only marker type is FragmentStart, which has zero width,
and is treated as inert by the word wrapping system.

This involves a lot of boilerplate changes to destructure Str(thing)
to a TaggedString, and vice versa. The only 'interesting' part of the
change is the 'else' clause in WrappedBlock::flush_word() which passes
through any TaggedLineElement other than Str directly to the output
wrapped line.

I've also changed the signature of TaggedLine::iter() so that it
exposes the new TaggedLineElement element type in full, so html2term
can see the fragment start markers when it iterates over the word-
wrapped version of the document. But it doesn't actually do anything
with them yet.

(This diff is probably clearer to view with whitespace ignored,
because it reindents a large block in flush_word().)