Fragment support: follow intra-document hyperlinks in html2term.
This is the payoff from the previous two commits. html2term now scans
the formatted document for FragmentStart markers and sets up a HashMap
from fragment name to document (x,y) coordinates. Pressing Return when
the cursor is on a hyperlink, if the link target refers to a fragment
within the same document (e.g. bare <a href="#foo"> with no file
name), will look up the fragment id in that map and jump to its
location.

Hence, you can now use html2term to view an HTML document containing
internal hyperlinks (e.g. table of contents, or footnotes), and use
those links to navigate conveniently around.

I haven't added any support for following links _between_ files, or
for remembering the location history to implement a back button. This
is just the very basics of being able to follow links at all.

Also, the current policy on scrolling doesn't work terribly well with
the jumps through the document made by link-following. Currently we
scroll by the minimum amount necessary to put the new cursor position
within the visible screen. So if you jump from the table of contents
down to a section heading, the heading and the cursor both end up on
the _bottom_ line of the screen, so you have to hit Page Down before
you can start reading the actual body text of the section you were
interested in. A better heuristic would be nice.