html2text: add a --literal option to omit markup.
This is a useful option if the input HTML file has been generated by
an MUA which automatically inserts hyperlinks around parts of the
message that look like URLs or email addresses, and you want to
extract literal text from the message such as a shell transcript.

My technique is a bit bodgy. In text_renderer.rs I thought it was
cleaner to make a separate TrivialDecorator type than to complicate
the existing PlainDecorator with a boolean flag and endless ifs. But
then in html2text.rs I couldn't find a way to initialise a variable to
one of TrivialDecorator and PlainDecorator depending on the command-
line flag, so I used an extra layer of wrapper function, which surely
isn't the best way. But it seems to work.