Add parse function as a new entry point
Previously, there were three different entry points for html2text:
from_read, from_read_with_decorator and from_read_rich.  This had two
drawbacks:

Firstly, the user always had to parse and render the HTML at the same
time.  But there are use cases where the user wants to render the same
HTML document using different widths or decorators.  This involved
parsing the document multiple times.

Secondly, it was impossible to create rich text with a custom decorator.
This would have required a new function from_read_rich_with_decorator.

This patch introduces a new entry point for the pubic API, the parse
function.  It reads and parses the HTML document and returns a
RenderTree object.  This object can be rendered into a RenderedText
object using a custom decorator or the plain or rich decorators.  The
RendererText object can be converted to a string or to a vector of
annotated lines.