Keep annotations in TextRenderer::into_lines
The TextRenderer::into_lines method calls the TextDecorator::finalise
method and appends the returned lines to the rendered lines, re-wrapping
them if required.  Previously, it would concatenate all strings into
lines and append one TaggedString without annotations per line.

With this patch, we keep the original tagged strings and just split them
if required.  This means that the annotations are no longer stripped
from the lines returned by finalise.  We also add a new test case that
checks that these annotations are present.