Add into_tagged_strings method to TaggedLine
In commit 17940794eacceb36679ca04e54573e32ae8906ca, we added the
tagged_strings method to TaggedLine that returns an iterator over the
tagged strings.  In this patch, we also add a into_tagged_strings method
that takes ownership of the tagged strings.  This is useful when
converting a TaggedLine, for example in the TaggedLine::width method or
in external crates that process the TaggedLine before rendering it.