Fix: ignore wrapping for links
Arguably, wrapping links is not particularly helpful. Wrapping to a given width is
sensible for elements where html2text must remain in control of the visual
appearance, but the links that come after the main content are a block of text
where the width to which they are wrapped is not really relevant. Worse, wrapping
means off-the-shelf tools which could easily extract all the links from the bottom
of the plaintext output no longer work, but must deal with linebreaks inside of
links. By turning off wrapping specifically of links, these tools work as expected.