Fix nits in documentation.
The doc comment for use_doc_css had an unescaped <style>, which was
emitted unchanged into the HTML by rustdoc and ate the whole of the
rest of the HTML output file. I've escaped it so that it's printed
literally.

TextDecorator's top-level documentation had a paragraph ending with an
unfinished sentence. I've finished it as best I could, although of
course I may have guessed wrong!

Several decorator methods had the word 'an' before a consonant. I
suspect it was copy-paste error from 'after an em' just above.