introduce rendering option to extract raw text
When the raw option is set, subrenderer for cells in tables use the full
available width and ensure text within a cell is next to eachother in the
rendered text.
Raw mode additionally does not draw table boundaries.

The implementation of this must switch the iteration order from first
iterating over same lines of adjacent cells to first iterating over the cells and
then over every line of that cell.