Handle some in-the-wild invalid colours.
I have had an e-mail from a large institution with something like
`<table bgcolor="012345">`. The bgcolor attribute is no longer
supported, and as far as I can find was never supported with just hex
digits and no hash - but both Firefox and Chromium accept it and parts
of the e-mail were white-on-white without.