Remove fn parse() from public API and replace with pub fn do_parse() method on Config
The parse function is an odd one in the current API, other functions
can be separated cleanly into functions that are convenient and functions
that need to be configured. The parse function does not advertise what
configuration it uses, and will silently drop css information, as it parses
with config::plain() as default configuration. This can lead to the following
misuse of the public API

let render_tree = parse(html)?;
customcfg.render_to_string(render_tree, 80)?;