Fix missing FragmentStarts.
This could happen when the `id=` attribute was on an element which did
not cause any text to be output, and the fragments were never flushed to
the output as they were considered empty.

We now keep track of pending fragments and output them to the start of
the next text line being output.