Allow for syntax highlighting `<pre>` blocks (with "css_ext" feature)
Add `Config::register_highlighter()`, which can add a
syntax highlighting function associated with a language name. This can
be used with a CSS declaration `x-syntax: name`.  The highlighter
is passed a raw string containing all the text inside the `<pre>`, not
including any elements, and returns a list of styled spans.  Foreground
and background colours are currently supported.

The CSS extensions are no longer available in document/author CSS, since
we're likely to want to view untrusted HTML.  In an application which
wants to give HTML the ability to do so, it can be done via a class and
an extra agent rule referring to that class.

Add `--syntax` option to the html2text example to register a `rs` (rust)
highlighter as an example, which can be used with `--agent-css='pre.rust
{ x-syntax: rs; }` (`--agent-css` is also new).

Fixes #36.